+++ infra/resources/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @infra/resources
 
+## 0.2.0
+
+### Minor Changes
+
+- 1f38d3d: [CES-540] Create a DB in the existing Cosmos account
+
+## 0.1.1
+
+### Patch Changes
+
+- 0423350: Update CostCenter value
+
 ## 0.1.0
 
 ### Minor Changes